Fixed stuff

+++ tools/find_all_files.py
@@ -11,17 +11,27 @@ def find_files_by_extension(source_dir, target_dir, extensions):
                 file_path = os.path.join(root, file)
                 file_paths.append(file_path)
 
+    errors = 0
+    skipped = 0
+    copied = 0
+
     for file_path in tqdm.tqdm(file_paths, desc="Moving files"):
+        target_path = os.path.join(target_dir, os.path.basename(file_path))
+        if os.path.exists(target_path):
+            skipped += 1
+            continue
         try:
             shutil.move(file_path, target_dir)
-            print(f"Moved '{file_path}' to '{target_dir}'")
+            copied += 1
         except Exception as e:
             print(f"Error moving '{file_path}': {e}")
+            errors += 1
+    
+    print(f"Files copied: {copied}, Files skipped: {skipped}, Errors: {errors}")
 
 if __name__ == "__main__":
-    source_dir = r"D:\Data\Rubber"
+    source_dir = r"C:\Users\bertr\Downloads\winpe"
     target_dir = r"D:\Data\temp"
     extensions = ['.zip', '.exe', '.gz', '.bz2', '.xz', '.7z', '.rar']
 
     find_files_by_extension(source_dir, target_dir, extensions)
-
